Sera ACU regen field rework
As per call changed the slots of regen field and ras around (I know there is an open PR for it but it seems to be missing a few lines of upgrade VFX code.
Rebalanced to have a regen ceiling based on a tech level as opposed to a shared one.
Regen field ceilings:
* T1 - 10
* T2 - 15
* T3 - 25
* T4 - 40

Advanced regen field ceilings:
* T1 - 20
* T2 - 50
* T3 - 120
* T4 - 240